Move structs to models
