+++ src/models/models.go
@@ -1,0 +1,15 @@
+package models
+
+type User struct {
+	UserId       int64
+	Username     string
+	Email        string
+	PasswordHash string
+}
+
+type Message struct {
+	Author  User
+	PubDate int64  // The publish timestamp as UNIX
+	Text    string // The message itself
+	Flagged bool
+}
